Move 'ticker.ts' into 'channels' sub-directory to align with the Channels guide.
