+++ examples/channels/channels/ticker.ts
