+++ examples/channels/public/index.js
@@ -1,15 +1,15 @@
+const eventSource = new EventSource("/sse");
+
 const count = document.createElement("pre");
 count.innerText = "No count just yet.";
 document.body.appendChild(count);
 
-const activeSessions = document.createElement("pre");
-document.body.appendChild(activeSessions);
-
-const eventSource = new EventSource("/sse");
-
 eventSource.addEventListener("tick", ({data}) => {
 	count.innerText = `The clock has ticked! The count is now ${data}.`;
 });
+
+const activeSessions = document.createElement("pre");
+document.body.appendChild(activeSessions);
 
 eventSource.addEventListener("session-count", ({data}) => {
 	activeSessions.innerText = `There are ${data} person(s) watching this pointless number.`;
